+++ uCOS-II_Wrapper/uCOS-II/os_cfg.h
@@ -137,7 +137,7 @@
 #endif
 #define OS_Q_ACCEPT_EN            0u   /*     Include code for OSQAccept()                             */
 #define OS_Q_DEL_EN               1u   /*     Include code for OSQDel()                                */
-#define OS_Q_FLUSH_EN             0u   /*     Include code for OSQFlush()                              */
+#define OS_Q_FLUSH_EN             1u   /*     Include code for OSQFlush()                              */
 #define OS_Q_PEND_ABORT_EN        0u   /*     Include code for OSQPendAbort()                          */
 #define OS_Q_POST_EN              1u   /*     Include code for OSQPost()                               */
 #define OS_Q_POST_FRONT_EN        1u   /*     Include code for OSQPostFront()                          */